+++ nope/lib/views/page/form.php
@@ -4,8 +4,8 @@
       <div class="col col-md-9">
         <div class="panel panel-default">
           <div class="panel-body">
-            <div class="form-group">
-              <label>Title</label>
+            <div class="form-group" ng-class="{'has-error':(!contentForm.title.$valid && contentForm.title.$touched)}">
+              <label class="control-label">Title</label>
               <input type="text" name="title" class="form-control input-lg" ng-model="content.title" required />
             </div>
             <div class="form-group">
@@ -32,29 +32,27 @@
               <div class="preview-image" ng-show="content.cover" ng-style="{backgroundImage:'url('+content.cover.preview.thumb+')'}">
                 <a href="" ng-click="content.cover=null" class="btn btn-danger btn-xs"><i class="fa fa-times-circle"></i></a>
               </div>
-              <nope-model model="Media" ng-model="content.cover" preview="icon" multiple="false" label="Add featured image"></nope-model>
+              <nope-model model="Media" ng-model="content.cover" multiple="false" label="Add featured image"></nope-model>
             </div>
-            <div class="form-group content-author">
-
+            <div class="form-group" ng-class="{'has-error':(!contentForm.slug.$valid && contentForm.slug.$touched)}">
+              <label class="control-label">Slug</label>
+              <input type="text" name="slug" class="form-control input-sm" ng-model="content.slug" required ng-pattern="/^[a-zA-Z0-9-_\/]+$/" ng-trim="false" />
             </div>
-            <div class="form-group">
-              <label>Slug</label>
-              <input type="text" name="slug" class="form-control" ng-model="content.slug" required ng-pattern="/^[a-zA-Z0-9-_\/]+$/" ng-trim="false" />
-            </div>
-            <div class="form-group">
-              <label>Status</label>
-              <select class="form-control" ng-model="content.status" required>
+            <div class="form-group content-status" ng-class="{'has-error':(!contentForm.status.$valid && contentForm.status.$touched)}">
+              <label class="control-label">Status</label>
+              <nope-publishing ng-model="content"></nope-publishing>
+              <select class="form-control input-sm" ng-model="content.status" required ng-blur="getRealStatus()">
                 <option value="draft">Draft</option>
                 <option value="published">Published</option>
               </select>
             </div>
             <div class="form-group">
               <label>Start publishing date</label>
-              <input type="text" name="title" class="form-control" ng-model="content.startPublishingDate" placeholder="yyyy-mm-dd hh:mm:ss" />
+              <input type="text" name="startPublishingDate" class="form-control input-sm" ng-model="content.startPublishingDate" placeholder="yyyy-mm-dd hh:mm:ss" ng-blur="getRealStatus()" />
             </div>
             <div class="form-group">
               <label>End publishing date</label>
-              <input type="text" name="title" class="form-control" ng-model="content.endPublishingDate" placeholder="yyyy-mm-dd hh:mm:ss" />
+              <input type="text" name="endPublishingDate" class="form-control input-sm" ng-model="content.endPublishingDate" placeholder="yyyy-mm-dd hh:mm:ss" ng-blur="getRealStatus()" />
             </div>
             <div class="form-group">
               <button class="btn btn-block" ng-disabled="contentForm.$invalid" ng-class="{'btn-success':!contentForm.$invalid}">Save</button>